server/requirements.txt
# ClinicGuard-AI Requirements
# Note: This application runs a single FastAPI service that handles all AI functionality
# (Whisper transcription, LLaMA generation, and TTS) directly without microservices.

fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
twilio==8.12.0
python-multipart==0.0.9
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0
tenacity==8.2.3
python-socketio==5.11.1
websockets==12.0
email-validator==2.1.0.post1
llama-cpp-python==0.2.23
openai-whisper==20231117
numpy==1.26.3
pyttsx3==2.90
pydantic==2.10.3
accelerate==1.0.1
datasets==2.14.1
huggingface-hub==0.23.3 
torchaudio==2.3.1
typing-extensions==4.12.2
torch==2.3.1
transformers==4.38.2
openai==1.12.0